indigo_hack/requirements.txt

Django==2.2.7
django-cors-headers==3.1.1
django-rest-auth==0.9.5
django-shell-plus==1.1.7
django-storages==1.9.1
django-suit==0.2.28
djangorestframework==3.10.3
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.3.0
django-log-request-id==1.6.0
django-requestlogs==0.2.1
